+++ internal/controller/operator/factory/vmagent/podscrape.go
@@ -36,14 +36,18 @@ func generatePodScrapeConfig(
 
 	selectedNamespaces := getNamespacesFromNamespaceSelector(&sc.Spec.NamespaceSelector, sc.Namespace, se.IgnoreNamespaceSelectors)
 	if ep.AttachMetadata.Node == nil && sc.Spec.AttachMetadata.Node != nil {
-		ep.AttachMetadata = sc.Spec.AttachMetadata
+		ep.AttachMetadata.Node = sc.Spec.AttachMetadata.Node
 	}
+	if ep.AttachMetadata.Namespace == nil && sc.Spec.AttachMetadata.Namespace != nil {
+		ep.AttachMetadata.Namespace = sc.Spec.AttachMetadata.Namespace
+	}
+
 	k8sSDOpts := generateK8SSDConfigOptions{
 		namespaces:         selectedNamespaces,
 		shouldAddSelectors: cr.Spec.EnableKubernetesAPISelectors,
 		selectors:          sc.Spec.Selector,
 		apiServerConfig:    apiserverConfig,
-		role:               kubernetesSDRolePod,
+		role:               k8sSDRolePod,
 		attachMetadata:     &ep.AttachMetadata,
 		namespace:          sc.Namespace,
 	}